backend/requirements.txt
# Core API
fastapi
uvicorn

# Environment variables
python-dotenv
python-multipart

# LLM (Mistral via Hugging Face)
langchain
langchain-huggingface
huggingface-hub
langchain-community
langchain-classic

# Utilities
pydantic
